// Desktop/pfc/project/keybop-scraper/src/app/interfaces/ProductDetailInterface.ts
import { ProductListInterface } from "./ProductListInterface";

export interface ProductDetailInterface {
  name:                string;
  price:               string;
  short_description:   string;
  images:              string;
  rating:              string;
  number_of_reviews:   string;
  variants:            string;
  product_description: string;
  sales_rank:          string;
  link_to_all_reviews: string;
  id:                  number;
}